compiling but page fault:(

+++ kernel/IDT/time.c
@@ -1,17 +1,17 @@
 #include "time.h"
 
- uint64_t g_seconds_count = 0;
- uint64_t g_ms_counter = 0;
+ uint32_t g_seconds_count = 0;
+ uint32_t g_ms_counter = 0;
  callback_function g_scheduler;
 
 void print_time_seconds();
-void sleep(uint64_t sleep_ms);
+void sleep(uint32_t sleep_ms);
 void set_scheduler(callback_function scheduler);
 
 
 void time_handler(registers_t* registers)
 {
-    static uint64_t i = 0;
+    static uint32_t i = 0;
     g_ms_counter += 55;
     g_seconds_count = g_ms_counter / 1000;
 
@@ -32,10 +32,10 @@ void print_time_seconds()
     putc('\n');   
 }
 
-void sleep(uint64_t sleep_ms)
+void sleep(uint32_t sleep_ms)
 {
     // get curr miliseconds time 
-    uint64_t curr_ms = g_ms_counter;
+    uint32_t curr_ms = g_ms_counter;
 
     // loop untile enough seconds waited
     while(curr_ms + sleep_ms > g_ms_counter){}
