+++ kernel/IDT/time.h
@@ -24,7 +24,7 @@ void print_time_seconds();
     this function sleeps for specific amount of miliseconds
     @param sleep_ms: the wanted sleep time
 */
-void sleep(uint64_t sleep_ms);
+void sleep(uint32_t sleep_ms);
 
 /*
     This function sets the scheduler function to be called every 200 miliseconds
